changed the health and the power from float and double into int, removed un-necessary includes

+++ Unit Armies/ES.cpp
@@ -1,6 +1,6 @@
 #include "ES.h"
 
-ES::ES(int id, int tj, double health, float power, int attackCap,
+ES::ES(int id, int tj, int health, int power, int attackCap,
 	std::string type):Unit(id,tj,health,power, attackCap, type)
 {
