+++ Unit Armies/ET.h
@@ -5,7 +5,7 @@ class ET :public Unit {
 
 
 public:
-	ET(int id = 0, int tj = 0, double health = 0, float power = 0, int attackCap = 0,
+	ET(int id = 0, int tj = 0, int health = 0, int power = 0, int attackCap = 0,
 		std::string type = " ");
 
 	void attack();
