+++ Unit Armies/Unit.h
@@ -16,15 +16,15 @@ protected:
 
 	int ID;
 	int Tj;
-	double health;
-	float power;
+	int health;
+	int power;
 	int attackCapacity;
 	std::string type;
 	
 
 public:
 
-	Unit(int id = 0, int tj = 0, double health = 0, float power = 0, int attackCap = 0,
+	Unit(int id = 0, int tj = 0, int health = 0, int power = 0, int attackCap = 0,
 		std::string type = NULL) :ID(id), Tj(tj), health(health), power(power),
 		attackCapacity(attackCap), type(type){};
 	
@@ -32,8 +32,8 @@ public:
 
 	int GetID() const { return ID; };
 	int GetTj() const { return Tj; };
-	double GetHealth() const { return health; };
-	float GetPower() const { return power; };
+	int GetHealth() const { return health; };
+	int GetPower() const { return power; };
 	int GetAttackCap() const { return attackCapacity; };
 	std::string GetType() const { return type; };
